+++ App.js
@@ -4,13 +4,13 @@ import { TouchableOpacity, Image } from "react-native";
 import "react-native-gesture-handler";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
-import Splash from "./src/components/splash.js";
+import Splash from "./src/screens/splash.js";
 import LoginSignup from "./src/components/session/loginSignup.js";
-import IntroVideo from "./src/components/video.js";
-import Storytime from "./src/components/storytime.js";
-import MilkMilkMilk from "./src/components/milkMilkMilk";
-import Profile from "./src/components/profile.js";
-import Mindfulness from "./src/components/mindfulnessStack";
+import IntroVideo from "./src/screens/video.js";
+import Storytime from "./src/screens/storytime.js";
+import MilkMilkMilk from "./src/screens/milkMilkMilk";
+import Profile from "./src/screens/profile.js";
+import Mindfulness from "./src/screens/mindfulnessStack";
 import Counting from "./src/screens/Counting";
 import { Provider } from "react-redux";
 import configureStore from "./src/store/store";
@@ -21,18 +21,18 @@ import Activities from "./src/screens/Activities";
 import BoxBreathing from "./src/screens/boxBreathing";
 import ChatPlaceholder from "./src/screens/chatPlaceholder";
 import FeelingDictionary from "./src/screens/FeelingDictionary";
-import kpi from "./src/components/kpi";
+import kpi from "./src/screens/kpi";
 import { navigationRef } from "./src/components/RootNavigation";
 import * as RootNavigation from "./src/components/RootNavigation";
-import FiveFourThreeTwoOneTech from "./src/components/FiveFourThreeTwoOneTech.js";
+import FiveFourThreeTwoOneTech from "./src/screens/FiveFourThreeTwoOneTech.js";
 import CountingPrompt from "./src/screens/CountingPrompt.js";
 import CountingSelection from "./src/screens/CountingSelection.js";
 import { useFonts } from "expo-font";
 import Adventure from "./src/screens/Adventure.js";
 import AdventureLocation from "./src/screens/AdventureLocation.js";
 import AdventureLocationSeeAll from "./src/screens/AdventureLocationSeeAll.js";
-import FlatActivities from "./src/components/FlatActivities.js";
-import IntroActivity from "./src/components/IntroActivity.js";
+import FlatActivities from "./src/screens/FlatActivities.js";
+import IntroActivity from "./src/screens/IntroActivity.js";
 
 const Stack = createStackNavigator();
 const store = configureStore();
